Add type signatures for URI::HTTP and URI::HTTPS
